Fix PHPStan errors: type mismatches, regenerate baseline
- UserController: pass media_id int instead of WP_Post to prepare_item
- MediaListPage: cast ceil() to int for pagination
- Regenerated baseline for level 5 (0 errors)

+++ includes/Admin/MediaListPage.php
@@ -84,7 +84,7 @@ class MediaListPage {
 			$total = (int) $wpdb->get_var( $count_sql ); // phpcs:ignore WordPress.DB.DirectDatabaseQuery, WordPress.DB.PreparedSQL.NotPrepared
 		}
 
-		$total_pages = ceil( $total / $per_page );
+		$total_pages = (int) ceil( $total / $per_page );
 
 		// Fetch rows.
 		$orderby    = 'created_at';
